Varios cambios: ahora el jugador guarda una referencia al tablero (no es mas un Singleton). Se agrega el metodo venderBarrio a la clase Jugador. El juego es un Singleton y crea al Tablero y a los jugadores. Se agrega llamado a actualizarCasillero en algunos encasillables.

+++ src/fiuba/algo3/tp2/algopoly/model/Dinero.java
@@ -4,7 +4,7 @@ public class Dinero {
 
     private double cantidad;
 
-    public Dinero(int cantidad) {
+    public Dinero(double cantidad) {
         this.cantidad = cantidad;
     }
